Update: widget_test 
Created a copy
